requirements.txt
asgiref==3.3.1
Django==3.1.4
django-crispy-forms==1.10.0
djangorestframework==3.12.2
Jinja2==2.11.2
MarkupSafe==1.1.1
modeltranslation==0.25
Pillow==8.0.1
pytz==2020.4
sqlparse==0.4.1
